@@ -20,5 +20,5 @@ Gem::Specification.new do |spec|
 
   spec.files = Dir["{app,config,db,lib}/**/*", "MIT-LICENSE", "Rakefile", "README.md"]
 
-  spec.add_dependency "rails", ">= 4.2.0"
+  spec.add_dependency "rails", "~> 6.1.4", ">= 6.1.4.1"
 end
