@@ -1,6 +1,6 @@
-class CreateDoAddonConnectorSsoEvents < ActiveRecord::Migration
+class CreateDoAddonConnectorSsoEvents < ActiveRecord::Migration[6.1]
   def change
-    create_table :do_addon_connector_sso_events do |t|
+    create_table :do_addon_connector_sso_events, id: :uuid do |t|
       t.string :resource_uuid
       t.string :resource_token
       t.integer :timestamp
